+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## 7.4.0 - 2026-07-31
+* Introduced `TaskTemplates API`.
+* `variables` field was deprecated in the `Task` interface.
+
 ## 7.3.0 - 2026-07-15
 * Introduced `DecisionStrategyExecutionMode` enum.
 * Added `executionMode` (optional `DecisionStrategyExecutionMode`) to the `MakeDecisionParams` interface.